+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **v6.0.0 (unreleased):**
 
 * Removed compatibility patches for nbconvert < 6 per [#777](https://github.com/ucbds-infra/otter-grader/issues/777)
+* Updated Otter Export to throw an error if nbconvert<6.0.0 is found
 
 **v5.6.0:**